+++ flex/trader/query.ts
@@ -3,11 +3,27 @@ import { Market } from "../market";
 import { Token } from "../token";
 import { WalletInfo, walletInfoFromApi } from "../client";
 import { OrderInfo, TradeInfo } from "./types";
+import { uint256 } from "../web3";
+
+function orderInfoFromApi(result: any): OrderInfo {
+    return {
+        orderId: result.orderId,
+        traderId: result.userId,
+        price: result.price,
+        priceNum: result.priceNum,
+        priceScale: result.priceScale,
+        amountProcessed: result.amountProcessed,
+        amountLeft: result.amountLeft,
+        side: result.side,
+        finishTime: result.finishTime,
+        pair: result.pair,
+    };
+}
 
 /** @internal */
 export async function queryOrders(flex: Flex, trader: string): Promise<OrderInfo[]> {
     const result = await flex.query(`
-            userOrders(userId:"0x${trader}") {
+            userOrders(userId:"${uint256(trader)}") {
                 pair { ${Market.queryFields()} }
                 side
                 price
@@ -17,15 +33,31 @@ export async function queryOrders(flex: Flex, trader: string): Promise<OrderInfo
                 userId
                 amountProcessed
                 amountLeft
+                finishTime
             }
         `);
-    return result.userOrders;
+    return result.userOrders.map(orderInfoFromApi);
+}
+
+function tradeFromApi(result: any): TradeInfo {
+    return {
+        pair: result.pair,
+        price: result.price,
+        amount: result.amount,
+        time: result.time,
+        side: result.side,
+        liquidity: result.liquidity,
+        fees: result.fees,
+        feesToken: result.feesToken,
+        userOrderId: result.userOrderId,
+        cursor: result.cursor,
+    };
 }
 
 /** @internal */
 export async function queryTrades(flex: Flex, trader: string): Promise<TradeInfo[]> {
     const result = await flex.query(`
-        userTrades(userId:"0x${trader}") {
+        userTrades(userId:"${uint256(trader)}") {
             pair { ${Market.queryFields()} }
             price
             amount
@@ -38,21 +70,20 @@ export async function queryTrades(flex: Flex, trader: string): Promise<TradeInfo
             cursor
         }
     `);
-    return result.userTrades;
+    return result.userTrades.map(tradeFromApi);
 }
 
 export type QueryWalletsOptions = {
-
-    clientAddress: string,
+    clientAddress: string;
     /**
      * Trader ID as uint256 hex string
      */
-    traderId?: string,
+    traderId?: string;
     /**
      * Token DEX Wrapper address
      */
-    token?: string,
-}
+    token?: string;
+};
 
 /** @internal */
 export async function queryWallets(
@@ -62,7 +93,7 @@ export async function queryWallets(
     const result = await flex.query(`
         wallets(
             clientAddress: "${options.clientAddress}"
-            ${options.traderId ? `userId: "0x${options.traderId}"` : ""}
+            ${options.traderId ? `userId: "${uint256(options.traderId)}"` : ""}
             ${options.token ? `token: "${options.token}",` : ""}
         ) {
             address